Split related functions to individual .py files
