+++ TEST/test.py
@@ -1,0 +1,18 @@
+import pandas as pd
+import datetime as dt
+import cufflinks as cf
+import matplotlib.pyplot as plt
+import datetime as dt
+from calendar import day_name
+import PySimpleGUI as sg
+from date_gen import dateGEN
+
+# import constant
+
+
+
+### FUNCTIONS ###
+
+dateGEN(2020, 30)
+
+print('code compl')
